llm/requirements.txt
Flask==2.3.2
flask-cors==4.0.0
python-dotenv==1.0.1
supabase==1.2.0
sentence-transformers==4.1.0
google-generativeai==0.3.2
torch>=1.9.0
transformers>=4.30.0
numpy>=1.21.0
